Recompute CMGDB in the replay tier instead of re-rendering it
The replay tier re-rendered saved Morse data and never called CMGDB, so it
could not detect a change in the computation it exists to check. It now
reuses the trained networks, scalers and datasets and recomputes everything
downstream of them: box map, Morse graph, Conley indices, coarsenings,
regions of attraction, sampled residual/tolerance, figures.

Three changes make that possible without touching replay_sources/:

- morse_graph.run gains out_dir, separating where the model is read from
  where CMGDB writes. The morse stage is exempt from the read-only block
  when replay routing is active, and render/metrics consume the freshly
  written graph rather than re-rendering the preserved one.
- cmgdb.bounds_from_run_log reuses each run's own recorded latent rectangle
  from its mg_params_log.txt. Three of four families cannot re-infer bounds
  (coral ships no val.csv; leslie_2gen's dataset was never committed), and
  recomputing on the published domain keeps any difference attributable to
  CMGDB rather than to a shifted rectangle. Loading the training data is now
  lazy, so pinned bounds need no dataset at all.
- Absent stored-transition CSVs skip with a warning instead of aborting, and
  results record stored_transitions_missing beside what was read. In every
  published run the residual maximum came from runtime-generated Sobol or
  decoder-guided samples, so a missing source cannot lower a maximum it
  never attained.

Verified against the frozen reference results, with device pinned to CPU:

  coral seed 16      Morse sets bit-identical
  leslie3d_example1  all 122,346 boxes exact; merged Conley index
                     (0, x+1, 0) at 322 cells; closures and panels exact
  leslie_2gen        5 nodes, indices and 4 edges exact; +4 boxes in
                     132,180, a superset with nothing lost
  chafee_infante d=2 fully exact, including all 3,086 exact-RoA cell labels

Float32 box-map evaluation is device-dependent: MPS shifted 12 boundary
cells on leslie3d, so the CMGDB steps pin device: cpu.

Also fixed, all of which blocked a reader:

- cmgdb_roa shifted np.uint64 by a Python int, which works under NumPy 2's
  NEP 50 but raises under NumPy 1.x, on the exact-RoA path every 2-D latent
  example uses. Verified behaviour-preserving against 3,086 published labels.
- paper_figures_render could never succeed; it reads a bundle layout only
  leslie3d_example1_package_bundle.py produces, and that step was absent
  from the plan. Replaced by the packaging step, which renders the panels.
- psutil was imported by the Chafee d=3 controller but declared nowhere,
  breaking test collection.
- data_dir on the chafee and leslie3d configs pointed at trees that do not
  ship.
- Dependency floors were a snapshot, not a requirement: nothing reads a
  runtime version. Lowered to the oldest release carrying an API in use, and
  numpy is capped below 2.0 on macOS x86_64, where the last available torch
  (2.2.2) predates the numpy 2 ABI and aborts under it.

Steps whose inputs were never committed are marked as a blocked tier, never
selected, each naming what it needs. The Chafee d=2 residual row runs the
base seed only; the published row merges nine seeds, which REPRODUCING.md
had not documented.

+++ src/latentdynamics/analysis/cmgdb_roa.py
@@ -211,7 +211,11 @@ def compute_exact_roa(
             f"compute_exact_roa: {len(minimal_sorted)} minimal Morse nodes "
             f"exceeds the 64-bit reach-mask width"
         )
-    bit_of = {m: np.uint64(1) << i for i, m in enumerate(minimal_sorted)}
+    # Shift by a uint64, not a Python int: under NumPy 2's NEP 50 the int is a
+    # weak scalar that adopts uint64, but under NumPy 1.x uint64 and the int's
+    # int64 promote to float64, for which left_shift has no loop. Explicit
+    # uint64 on both sides behaves identically on either.
+    bit_of = {m: np.uint64(1) << np.uint64(i) for i, m in enumerate(minimal_sorted)}
     reach_mask = np.zeros(n_vertices, dtype=np.uint64)
     for minimal in minimal_sorted:
         targets = morse_cells.get(minimal, np.empty(0, dtype=np.int64))
